requirements.txt
# Core A2A dependencies
a2a-sdk>=0.3.0
protobuf

# Web server
uvicorn
starlette
fastapi

# LLM providers
google-adk
google-genai
langchain-google-genai
litellm
openai
anthropic

# Google Cloud dependencies
google-api-core
google-auth
googleapis-common-protos
grpcio

# Utilities
python-dotenv
httpx
pydantic
structlog

# Other dependencies
pyyaml
jinja2
typing-extensions
aiohttp  # Added for a2a_client.py